SummarizedExperiments for Julia

Quick start

Users may install this package from the GitHub repository through the usual process on the Pkg REPL:

add https://github.com/LTLA/MultiAssayExperiment.jl

And then:

Class definition

MultiAssayExperiments.MultiAssayExperiment — Type

The MultiAssayExperiment class is a Bioconductor container for multimodal studies. This is basically a list of SummarizedExperiment objects, each of which represents a particular experimental modality. A mapping table specifies the relationships between the columns of each SummarizedExperiment and a conceptual "sample", assuming that each sample has data for zero, one or multiple modalities. A sample can be defined as anything from a cell line culture to an individual patient, depending on the context.

This implementation makes a few changes from the original Bioconductor implementation. We do not consider the MultiAssayExperiment to contain any "columns", as this was unnecessarily confusing. The previous colData field has thus been renamed to sampledata, to reflect the fact that we are operating on samples. We are also much more relaxed about harmonization between the experiments, sample mapping, and sample data - or more specifically, we don't harmonize at all, allowing greater flexibility in storage and manipulation.

Constructors

MultiAssayExperiments.MultiAssayExperiment — Type

MultiAssayExperiment(experiments, sampledata, samplemap, metadata = Dict{String,Any}())

Creates a new MultiAssayExperiment from its components.

experiments should contain ordered pairs of experiment names and SummarizedExperiment objects. Each SummarizedExperiment may contain any number and identity for the rows. However, the column names must be non-nothing and unique within each object.

Each row of sampledata corresponds to a conceptual sample. The first column should be called name and contain the names of the samples in a Vector{String}. Sample names are arbitrary but should be unique. Any number and type of other columns may be provided, usually containing sample-level annotations.

The samplemap table is expected to have 3 Vector{String} columns - sample, experiment and colname - specifying the correspondence between each conceptual sample and the columns of a particular SummarizedExperiment. Values in the sample column will be cross-referenced to values in the name column of the sampledata; values in the experiment column will be cross-referenced to the keys of experiments; and the colname column will be cross-referenced to the column names of each SummarizedExperiment. Note that values in the columns need not have a 1:1 match to their cross-referenced target; any missing values in one or the other will be ignored in the methods. Rows of the table should be unique.

The metadata stores other annotations unrelated to the samples.

Examples

julia> using MultiAssayExperiments

julia> import SummarizedExperiments, DataStructures, DataFrames;

julia> exp = DataStructures.OrderedDict{String,SummarizedExperiments.SummarizedExperiment}();

julia> exp["foo"] = SummarizedExperiments.exampleobject(100, 2);

julia> exp["bar"] = SummarizedExperiments.exampleobject(50, 5);

julia> cd = DataFrames.DataFrame(
           name = ["Aaron", "Michael", "Jayaram", "Sebastien", "John"],
           disease = ["good", "bad", "good", "bad", "very bad"]
       );

julia> sm = DataFrames.DataFrame(
           sample = ["Aaron", "Michael", "Aaron", "Michael", "Jayaram", "Sebastien", "John"],
           experiment = ["foo", "foo", "bar", "bar", "bar", "bar", "bar"],
           colname = ["Patient1", "Patient2", "Patient1", "Patient2", "Patient3", "Patient4", "Patient5"]
       );

julia> using MultiAssayExperiments;

julia> out = MultiAssayExperiment(exp, cd, sm)
MultiAssayExperiment object
  experiments(2): foo bar
  sampledata(2): name disease
  metadata(0):

MultiAssayExperiments.MultiAssayExperiment — Method

MultiAssayExperiment(experiments)

Creates an MultiAssayExperiment object from a set of experiments. The per-sample column data and sample mapping is automatically created from the union of column names from all experiments.

Examples

julia> using MultiAssayExperiments

julia> import SummarizedExperiments, DataStructures

julia> exp = DataStructures.OrderedDict{String,SummarizedExperiments.SummarizedExperiment}();

julia> exp["foo"] = SummarizedExperiments.exampleobject(100, 10);

julia> exp["bar"] = SummarizedExperiments.exampleobject(50, 20);

julia> out = MultiAssayExperiment(exp)
MultiAssayExperiment object
  experiments(2): foo bar
  sampledata(1): name
  metadata(0):

MultiAssayExperiments.MultiAssayExperiment — Method

MultiAssayExperiment()

Creates an empty MultiAssayExperiment object.

Examples

julia> using MultiAssayExperiments

julia> MultiAssayExperiment()
MultiAssayExperiment object
  experiments(0):
  sampledata(1): name
  metadata(0):

Getters

MultiAssayExperiments.experiment — Method

experiment(x[, i]; sampledata = false)

Extract the specified SummarizedExperiment from a MultiAssayExperiment x. i may be a positive integer no greater than the number of experiments in x, or a string specifying the name of the desired experiment. If i is not specified, it defaults to the first experiment in x.

If sampledata = true, we attempt to add the sample data of x to the coldata of the returned SummarizedExperiment. This is done by subsetting sampledata(x) based on sample mapping to the columns of the returned SummarizedExperiment - see expandsampledata for more details. If there are columns in the sampledata(x) and the coldata of the SummarizedExperiment with the same name but different values, the former are omitted with a warning.

Note that, if sampledata = true, the returned SummarizedExperiment will be a copy of the relevant experiment in x. If false, the returned object will be a reference.

Examples

julia> using MultiAssayExperiments;

julia> x = exampleobject();

julia> experiment(x)
100x10 SummarizedExperiments.SummarizedExperiment
  assays(3): foo bar whee
  rownames: Gene1 Gene2 ... Gene99 Gene100
  rowdata(2): name Type
  colnames: foo1 foo2 ... foo9 foo10
  coldata(3): name Treatment Response
  metadata(1): version

julia> experiment(x, 1); # same result

julia> experiment(x, "foo");

julia> experiment(x, "foo", sampledata = true) # add sample data
100x10 SummarizedExperiments.SummarizedExperiment
  assays(3): foo bar whee
  rownames: Gene1 Gene2 ... Gene99 Gene100
  rowdata(2): name Type
  colnames: foo1 foo2 ... foo9 foo10
  coldata(4): name Treatment Response disease
  metadata(1): version

MultiAssayExperiments.experiments — Method

experiments(x)

Return an ordered dictionary containing all experiments in the MultiAssayExperiment x.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> collect(keys(experiments(x)))
2-element Vector{String}:
 "foo"
 "bar"

MultiAssayExperiments.sampledata — Method

sampledata(x, check = true)

Return a DataFrame containing the sample data in the MultiAssayExperiment x.

The returned object should contain name as the first column, containing a vector of unique strings. If check = true, the function will check the validity of the sample data before returning it.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> names(sampledata(x))
2-element Vector{String}:
 "name"
 "disease"

MultiAssayExperiments.samplemap — Method

samplemap(x)

Return an ordered dictionary containing the sample mapping from the MultiAssayExperiment x.

The returned object should contain the sample, experiment and colname columns in that order. Each column should contain a vector of strings, and rows should be unique. If check = true, the function will check the validity of the sample data before returning it.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> names(samplemap(x))
3-element Vector{String}:
 "sample"
 "experiment"
 "colname"

MultiAssayExperiments.metadata — Method

metadata(x)

Return a dictionary containing the metadata from the MultiAssayExperiment x.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> collect(keys(metadata(x)))
1-element Vector{String}:
 "version"

Setters

MultiAssayExperiments.setexperiment! — Method

setexperiment!(x[, i], value)

Set experiment i in MultiAssayExperiment x to the SummarizedExperiment value. This returns a reference to the modified x.

i may be a positive integer, in which case it should be no greater than the length of experiments(x). It may also be a string specifying a new or existing experiment in x. If omitted, we set the first experiment by default.

Examples

julia> using MultiAssayExperiments;

julia> x = exampleobject();

julia> size(experiment(x, 2))
(50, 8)

julia> val = experiment(x);

julia> setexperiment!(x, 2, val);

julia> size(experiment(x, 2))
(100, 10)

MultiAssayExperiments.setexperiments! — Method

setexperiments!(x, value)

Set the experiments in the MultiAssayExperiment x to the OrderedDict value. This returns a reference to the modified x.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> y = copy(experiments(x));

julia> delete!(y, "foo");

julia> setexperiments!(x, y);

julia> collect(keys(experiments(x)))
1-element Vector{String}:
 "bar"

MultiAssayExperiments.setsampledata! — Method

setsampledata!(x, value)

Set the sample data in the MultiAssayExperiment x to the DataFrame value.

The returned object should contain name as the first column, containing a vector of unique strings. If check = true, the function will check the validity of the sample data before returning it.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> sd = copy(sampledata(x));

julia> sd[!,"stuff"] = [rand() for i in 1:size(sd)[1]];

julia> setsampledata!(x, sd);

julia> names(sampledata(x))
3-element Vector{String}:
 "name"
 "disease"
 "stuff"

MultiAssayExperiments.setsamplemap! — Method

setsamplemap!(x, value)

Set the sample mapping in the MultiAssayExperiment x to a DataFrame value. This returns a reference to the modified x.

value should contain the sample, experiment and colname columns in that order. Each column should contain a vector of strings, and rows should be unique.

sample should (but is not required to) correspond to the names of samples in sampledata(x); experiment should (but is not required to) correspond to the keys of experiments(x); and colname should (but is not required to) correspond to the columns of the corresponding SummarizedExperiment in experiments(x). This correspondence is used for convenient subsetting and extraction in the various extract methods.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> y = samplemap(x)[1:10,:];

julia> setsamplemap!(x, y);

julia> size(samplemap(x))[1]
10

MultiAssayExperiments.setmetadata! — Method

setmetadata!(x, value)

Set the metadata of a MultiAssayExperiment x to a dictionary value. This returns a reference to the modified x.

Examples

julia> using MultiAssayExperiments

julia> x = exampleobject();

julia> meta = copy(metadata(x));

julia> meta["version"] = "0.2.0";

julia> setmetadata!(x, meta);

julia> metadata(x)["version"]
"0.2.0"

Miscellaneous

Base.copy — Method

copy(x::MultiAssayExperiment)

Return a copy of x, where all components are identically-same as those in x.

Examples

julia> using MultiAssayExperiments, DataFrames

julia> x = exampleobject();

julia> x2 = copy(x);

julia> setsampledata!(x2, DataFrame(name=["A", "B"]));

julia> size(sampledata(x))
(5, 2)

julia> size(sampledata(x2)) # Change to reference is only reflected in x2.
(2, 1)

julia> stuff = experiments(x);

julia> delete!(stuff, "bar");

julia> collect(keys(experiments(x2)))
1-element Array{String,1}:
 "foo"

Base.deepcopy — Method

deepcopy(x::MultiAssayExperiment)

Return a deep copy of x and all of its components.

Examples

julia> using MultiAssayExperiments, DataFrames

julia> x = exampleobject();

julia> x2 = deepcopy(x);

julia> insertcols!(sampledata(x), 2, "WHEE" => 1:5); # References now point to different objects.

julia> names(sampledata(x2))
2-element Array{String,1}:
 "name"
 "disease"

Base.show — Method

print(io::IO, x::MultiAssayExperiment)

Print a summary of x.

MultiAssayExperiments.exampleobject — Method

exampleobject()

Create an example MultiAssayExperiment object. This is to be used to improve the succinctness of examples and tests.

Examples

julia> using MultiAssayExperiments 

julia> x = exampleobject()
MultiAssayExperiment object
  experiments(2): foo bar
  sampledata(2): name disease
  metadata(1): version

Contact

This package is maintained by Aaron Lun (@LTLA). If you have bug reports or feature requests, please post them as issues at the GitHub repository.